utils/migrate-to-pg/requirements.txt
# postgres
psycopg2-binary==2.8.6
python-dotenv==0.14.0
